pyproject.toml
[project]
name = "text-to-image-cli"
version = "0.1.0"
description = "CLI wrapper for OpenRouter GPT-5.4 Image 2"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
